requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
greenlet==3.3.0
idna==3.11
playwright==1.57.0
pyee==13.0.0
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
